+++ src/pages/payment-callback/index.ts
@@ -1,0 +1,1 @@
+export { PaymentCallbackPage } from '#/pages/payment-callback/ui/PaymentCallbackPage'